Ukenytt_Frida/requirements.txt
tabula-py
pandas
requests
flask
waitress
jpype1
